refactor: 마이페이지 디자인 리팩터링
