+++ src/components/dashboard/index.ts
@@ -1,0 +1,5 @@
+export { default as ActivitySection } from './ActivitySection';
+export { default as ProfileSection } from './ProfileSection';
+export { default as StatsCard } from './StatsCard';
+export { default as AnalysisSection } from './AnalysisSection';
+export { default as BadgeSection } from './BadgeSection';
